bcommand: fix problems with multiple storages
Unfortunately, the storage abstraction layer introduced
some problems with storage handling when multiple storages
are defined.

The code that connects to the sd uses the storage stored in
ua->jcr->res.wstore, but not all code sets this to the currently
requested storage.

This patch fixes this problem by taking care that the
given storage is set as ua->jcr->res.wstore